+++ DarkCodex/Resources.cs
@@ -57,6 +57,7 @@ namespace DarkCodex
         {
             public static bool IsLoaded;
             public static List<BlueprintAbility> Ability = new();
+            public static List<BlueprintBuff> Buff = new();
             public static List<BlueprintFeature> Feature = new();
             public static List<BlueprintActivatableAbility> Activatable = new();
             public static List<BlueprintItem> Item = new();
@@ -263,6 +264,7 @@ namespace DarkCodex
 
                 Stopwatch timer = Stopwatch.StartNew();
                 Load(Ability, Path.Combine(Main.ModPath, "resources", "Ability.bin"));
+                Load(Buff, Path.Combine(Main.ModPath, "resources", "Buff.bin"));
                 Load(Feature, Path.Combine(Main.ModPath, "resources", "Feature.bin"));
                 Load(Activatable, Path.Combine(Main.ModPath, "resources", "Activatable.bin"));
                 Load(Item, Path.Combine(Main.ModPath, "resources", "Item.bin"));
@@ -277,6 +279,8 @@ namespace DarkCodex
             {
                 if (bp is BlueprintAbility ability)
                     Ability.Add(ability);
+                else if (bp is BlueprintBuff buff)
+                    Buff.Add(buff);
                 else if (bp is BlueprintFeature feature)
                     Feature.Add(feature);
                 else if (bp is BlueprintActivatableAbility activatable)
@@ -290,6 +294,7 @@ namespace DarkCodex
             public static void Clear()
             {
                 Ability.Clear();
+                Buff.Clear();
                 Feature.Clear();
                 Activatable.Clear();
                 Item.Clear();
@@ -309,6 +314,7 @@ namespace DarkCodex
 
                 // save blueprints guids to binary files
                 Save(Ability, Path.Combine(Main.ModPath, "resources", "Ability.bin"));
+                Save(Buff, Path.Combine(Main.ModPath, "resources", "Buff.bin"));
                 Save(Feature, Path.Combine(Main.ModPath, "resources", "Feature.bin"));
                 Save(Activatable, Path.Combine(Main.ModPath, "resources", "Activatable.bin"));
                 Save(Item, Path.Combine(Main.ModPath, "resources", "Item.bin"));